Changement dans les records (bonne utilisation de type) (#3)
* add tailwind as package

* add vitest for frontend testing

* remove useless classes from backend

* add frontend tests to CI

* Rename a bunch of stuff to more relevant functions

* Remove old node versions

* Add dependency path cache to the action

* Stub SignUpController

* Implement Professeur

* Comment out test

* Clean up du code de base pour commencer avec une base propre

* Transformation de la classe JWTAuthResponse en record

* Modification pour que les record prennent un id sous forme de int et non de Long,

---------

+++ src/main/java/com/lacouf/rsbjwt/service/dto/interfaceDTO/UserDto.java
@@ -3,7 +3,7 @@ package com.lacouf.rsbjwt.service.dto.interfaceDTO;
 import com.lacouf.rsbjwt.model.auth.Role;
 
 public interface UserDto {
-    Long id();
+    int id();
     String firstName();
     String lastName();
     String email();
